+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is manually updated. Please file an issue if something is missing.
 
+## Unreleased
+
+- **Feature**: Allow migrating Google Takeout metadata to EXIF ([#430](https://github.com/pulsejet/memories/issues/430))
+
 ## v4.12.1 (2023-03-15)
 
 - **Feature**: Load full image on zoom ([#266](https://github.com/pulsejet/memories/issues/266))